+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add `http.request.header.cookie` to fields scrubbed by cookie rules. ([#5408](https://github.com/getsentry/relay/pull/5408))
 - Map `request_id` to `trace_id` in vercel log drain transform. ([#5333](https://github.com/getsentry/relay/pull/5333))
 - Populate `sentry.platform` during OTLP ingestion. ([#5411](https://github.com/getsentry/relay/pull/5411))
+- Add db attribute normalization to V2 spans. ([#5399](https://github.com/getsentry/relay/pull/5399))
 
 **Internal**: